requirements.txt
python-dotenv==1.0.1
supabase==2.3.4
sendgrid==6.11.0
streamlit==1.31.1
langchain==0.1.7